# chore: replace deprecation eslint plugin in favor of ts-eslint/no-deprecated