--- a/packages/charts/react-charting/src/Styling.ts
+++ b/packages/charts/react-charting/src/Styling.ts
@@ -6,7 +6,7 @@ export {
   DefaultEffects,
   DefaultFontStyles,
   DefaultPalette,
-  // eslint-disable-next-line deprecation/deprecation
+  // eslint-disable-next-line @typescript-eslint/no-deprecated
   EdgeChromiumHighContrastSelector,
   FontClassNames,
   FontSizes,
@@ -39,11 +39,11 @@ export {
   createTheme,
   focusClear,
   fontFace,
-  // eslint-disable-next-line deprecation/deprecation
+  // eslint-disable-next-line @typescript-eslint/no-deprecated
   getEdgeChromiumNoHighContrastAdjustSelector,
   getFadedOverflowStyle,
   getFocusOutlineStyle,
-  // eslint-disable-next-line deprecation/deprecation
+  // eslint-disable-next-line @typescript-eslint/no-deprecated
   getFocusStyle,
   getGlobalClassNames,
   getHighContrastNoAdjustStyle,