--- a/packages/charts/react-charting/src/components/StackedBarChart/MultiStackedBarChart.base.tsx
+++ b/packages/charts/react-charting/src/components/StackedBarChart/MultiStackedBarChart.base.tsx
@@ -148,6 +148,7 @@ export class MultiStackedBarChartBase extends React.Component<IMultiStackedBarCh
             directionalHint={DirectionalHint.topAutoEdge}
             id={this._calloutId}
             onDismiss={this._closeCallout}
+            // eslint-disable-next-line @typescript-eslint/no-deprecated
             preventDismissOnLostFocus={true}
             /** Keep the callout updated with details of focused/hovered bar */
             shouldUpdateWhenHidden={true}