--- a/packages/charts/react-charting/src/utilities/SVGTooltipText.tsx
+++ b/packages/charts/react-charting/src/utilities/SVGTooltipText.tsx
@@ -252,7 +252,7 @@ export class SVGTooltipText
   };
 
   private _onTooltipKeyDown = (ev: React.KeyboardEvent<SVGElement>): void => {
-    // eslint-disable-next-line deprecation/deprecation
+    // eslint-disable-next-line @typescript-eslint/no-deprecated
     if ((ev.which === KeyCodes.escape || ev.ctrlKey) && this.state.isTooltipVisible) {
       this._hideTooltip();
       ev.stopPropagation();