--- a/packages/react-components/react-aria/library/src/button/index.ts
+++ b/packages/react-components/react-aria/library/src/button/index.ts
@@ -1,5 +1,5 @@
 export { useARIAButtonProps } from './useARIAButtonProps';
-// eslint-disable-next-line deprecation/deprecation
+// eslint-disable-next-line @typescript-eslint/no-deprecated
 export { useARIAButtonShorthand } from './useARIAButtonShorthand';
 export type {
   ARIAButtonAlteredProps,