--- a/packages/react-components/react-aria/library/src/index.ts
+++ b/packages/react-components/react-aria/library/src/index.ts
@@ -1,5 +1,5 @@
 export {
-  // eslint-disable-next-line deprecation/deprecation
+  // eslint-disable-next-line @typescript-eslint/no-deprecated
   useARIAButtonShorthand,
   useARIAButtonProps,
 } from './button/index';