--- a/packages/react-components/react-components/src/unstable/index.ts
+++ b/packages/react-components/react-components/src/unstable/index.ts
@@ -7,7 +7,7 @@
 // - use/consume `*-preview` packages directly for preview/unstable Fluent UI core controls early access
 // =====================================================================================================
 
-/* eslint-disable deprecation/deprecation */
+/* eslint-disable @typescript-eslint/no-deprecated */
 export {
   Alert,
   alertClassNames,
@@ -37,7 +37,7 @@ export type {
   InfoLabelSlots,
   InfoLabelState,
 } from '@fluentui/react-infobutton';
-/* eslint-enable deprecation/deprecation */
+/* eslint-enable @typescript-eslint/no-deprecated */
 
 export {
   Virtualizer,