--- a/packages/react-components/react-menu/library/src/Menu.ts
+++ b/packages/react-components/react-menu/library/src/Menu.ts
@@ -2,7 +2,7 @@ export type {
   MenuContextValues,
   MenuOpenChangeData,
   MenuOpenEvent,
-  // eslint-disable-next-line deprecation/deprecation
+  // eslint-disable-next-line @typescript-eslint/no-deprecated
   MenuOpenEvents,
   MenuProps,
   MenuSlots,