--- a/packages/react-components/react-menu/library/src/MenuList.ts
+++ b/packages/react-components/react-menu/library/src/MenuList.ts
@@ -5,7 +5,7 @@ export type {
   MenuListProps,
   MenuListSlots,
   MenuListState,
-  // eslint-disable-next-line deprecation/deprecation
+  // eslint-disable-next-line @typescript-eslint/no-deprecated
   UninitializedMenuListState,
 } from './components/MenuList/index';
 export {