--- a/packages/react-components/react-menu/library/src/components/MenuList/index.ts
+++ b/packages/react-components/react-menu/library/src/components/MenuList/index.ts
@@ -6,7 +6,7 @@ export type {
   MenuListProps,
   MenuListSlots,
   MenuListState,
-  // eslint-disable-next-line deprecation/deprecation
+  // eslint-disable-next-line @typescript-eslint/no-deprecated
   UninitializedMenuListState,
 } from './MenuList.types';
 export { renderMenuList_unstable } from './renderMenuList';