--- a/packages/react-components/react-menu/library/src/components/index.ts
+++ b/packages/react-components/react-menu/library/src/components/index.ts
@@ -13,7 +13,7 @@ export type {
   MenuListProps,
   MenuListSlots,
   MenuListState,
-  // eslint-disable-next-line deprecation/deprecation
+  // eslint-disable-next-line @typescript-eslint/no-deprecated
   UninitializedMenuListState,
 } from './MenuList/index';
 export {