--- a/packages/react-components/react-radio/library/src/contexts/index.ts
+++ b/packages/react-components/react-radio/library/src/contexts/index.ts
@@ -2,7 +2,7 @@ export {
   RadioGroupContext,
   RadioGroupProvider,
   useRadioGroupContextValue_unstable,
-  // eslint-disable-next-line deprecation/deprecation
+  // eslint-disable-next-line @typescript-eslint/no-deprecated
   useRadioGroupContext_unstable,
 } from './RadioGroupContext';
 export { useRadioGroupContextValues } from './useRadioGroupContextValues';