--- a/packages/react-components/react-switch/library/src/index.ts
+++ b/packages/react-components/react-switch/library/src/index.ts
@@ -1,7 +1,7 @@
 export {
   Switch,
   renderSwitch_unstable,
-  // eslint-disable-next-line deprecation/deprecation
+  // eslint-disable-next-line @typescript-eslint/no-deprecated
   switchClassName,
   switchClassNames,
   useSwitchStyles_unstable,