[eslint-plugin-deprecation](https://github.com/gund/eslint-plugin-deprecation) is archived and no longer maintained. It's recommended to migrate to [@typescript-eslint/no-deprecated before migrating to eslint v9.  ](https://github.com/gund/eslint-plugin-deprecation/issues/78#issuecomment-2416972936)

<!--
Thank you for submitting a pull request!

Please verify that:
* [ ] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [ ] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->

perf for rule tested locally (nx run react:lint --skip-nx-cache): 

| before       | after          |
| ---------- | ---------- |
|  5410 ms  | 4775 ms    |

~ 12% faster 🚀 


## Related Issue(s)

<!-- Please link the issue being fixed so it gets closed when this is merged. -->
Partially implements #30334