@Hotell 2025-01-10
finally this is happening, please measure the perf ( AFAIR the new plugin should be significantly faster ) ty